Dev to prod

+++ src/app/categories/services/categories.service.ts
@@ -1,17 +1,22 @@
 import { Injectable } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 
-import { getFiltersByUrl, getFindOptionsByFilters } from "../../shared/crud";
+import { ActiveOrderEntity } from "../../orders/entities";
+import { getFindOptionsByFilters } from "../../shared";
 import type { PaginationArgsDto } from "../../shared/dtos";
-import type { CreateCategoryDto, UpdateCategoryDto } from "../dtos";
+import { OrderStatusEnum } from "../../shared/enums";
+import type { CreateCategoryInput, UpdateCategoryInput } from "../dtos";
 import { CategoryEntity } from "../entities";
 
 @Injectable()
 export class CategoriesService {
-	private findRelations = ["file"];
-	private findOneRelations = ["file"];
+	private findRelations = ["file", "products"];
+	private findOneRelations = ["file", "products"];
 
-	constructor(@InjectRepository(CategoryEntity) private readonly _categoriesRepository) {}
+	constructor(
+		@InjectRepository(CategoryEntity) private readonly _categoriesRepository,
+		@InjectRepository(ActiveOrderEntity) private readonly _ordersRepository
+	) {}
 
 	async getCategory(id: string) {
 		return this._categoriesRepository.findOne({
@@ -20,9 +25,8 @@ export class CategoriesService {
 		});
 	}
 
-	async getCategories({ take, skip, filtersString }: PaginationArgsDto) {
-		const filters = getFiltersByUrl(filtersString);
-		const findOptions = getFindOptionsByFilters(filters) as any;
+	async getCategories({ take, skip, filtersArgs }: PaginationArgsDto) {
+		const findOptions = getFindOptionsByFilters(filtersArgs) as any;
 
 		const [data, count] = await this._categoriesRepository.findAndCount({
 			where: findOptions.where,
@@ -38,7 +42,7 @@ export class CategoriesService {
 		};
 	}
 
-	async createCategory(category: CreateCategoryDto): Promise<CategoryEntity> {
+	async createCategory(category: CreateCategoryInput): Promise<CategoryEntity> {
 		const savedCategory = await this._categoriesRepository.save({ ...category, place: { id: category.place } });
 
 		return this._categoriesRepository.findOne({
@@ -46,11 +50,35 @@ export class CategoriesService {
 		});
 	}
 
-	async updateCategory(id: string, category: UpdateCategoryDto): Promise<CategoryEntity> {
-		return this._categoriesRepository.save({ id, ...category });
+	async updateCategory(id: string, category: UpdateCategoryInput): Promise<CategoryEntity> {
+		await this._categoriesRepository.save({ id, ...category });
+
+		return this._categoriesRepository.findOne({
+			where: { id },
+			relations: this.findOneRelations
+		});
 	}
 
 	async deleteCategory(id: string): Promise<string> {
+		const orders: ActiveOrderEntity[] = await this._ordersRepository.find({
+			where: {
+				productsToOrders: {
+					product: {
+						category: { id }
+					}
+				}
+			},
+			relations: ["productsToOrders", "productsToOrders.product", "productsToOrders.product.category"]
+		});
+
+		const isActiveOrdersPresent = orders.some((el) => el.status !== OrderStatusEnum.CLOSED);
+
+		if (isActiveOrdersPresent) {
+			await this._categoriesRepository.save({ id, isHide: true });
+
+			return `Category is using in active order(s). Category will be hide for now`;
+		}
+
 		await this._categoriesRepository.delete(id);
 		return "DELETED";
 	}
