+++ src/app/halls/services/halls.service.ts
@@ -1,10 +1,11 @@
 import { Injectable } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
-import { Repository } from "typeorm";
 
-import { getFiltersByUrl, getFindOptionsByFilters } from "../../shared/crud";
+import { ActiveOrderEntity } from "../../orders/entities";
+import { getFindOptionsByFilters } from "../../shared";
 import type { PaginationArgsDto } from "../../shared/dtos";
-import type { CreateHallDto, UpdateHallDto } from "../dtos";
+import { OrderStatusEnum } from "../../shared/enums";
+import type { CreateHallInput, UpdateHallInput } from "../dtos";
 import { HallEntity } from "../entities";
 
 @Injectable()
@@ -12,7 +13,10 @@ export class HallsService {
 	private findRelations = ["place", "tables", "file"];
 	private findOneRelations = ["place", "tables", "file"];
 
-	constructor(@InjectRepository(HallEntity) private readonly _hallsRepository: Repository<HallEntity>) {}
+	constructor(
+		@InjectRepository(HallEntity) private readonly _hallsRepository,
+		@InjectRepository(ActiveOrderEntity) private readonly _ordersRepository
+	) {}
 
 	async getHall(id: string) {
 		return this._hallsRepository.findOne({
@@ -21,9 +25,8 @@ export class HallsService {
 		});
 	}
 
-	async getHalls({ take, skip, filtersString }: PaginationArgsDto) {
-		const filters = getFiltersByUrl(filtersString);
-		const findOptions = getFindOptionsByFilters(filters) as any;
+	async getHalls({ take, skip, filtersArgs }: PaginationArgsDto) {
+		const findOptions = getFindOptionsByFilters(filtersArgs) as any;
 
 		const [data, count] = await this._hallsRepository.findAndCount({
 			where: findOptions.where,
@@ -39,7 +42,7 @@ export class HallsService {
 		};
 	}
 
-	async createHall(hall: CreateHallDto): Promise<HallEntity> {
+	async createHall(hall: CreateHallInput): Promise<HallEntity> {
 		const savedHall = await this._hallsRepository.save({ ...hall, place: { id: hall.place } });
 
 		return this._hallsRepository.findOne({
@@ -47,11 +50,33 @@ export class HallsService {
 		});
 	}
 
-	async updateHall(id: string, hall: UpdateHallDto): Promise<HallEntity> {
-		return this._hallsRepository.save({ id, ...hall });
+	async updateHall(id: string, hall: UpdateHallInput): Promise<HallEntity> {
+		this._hallsRepository.save({ id, ...hall });
+
+		return this._hallsRepository.findOne({
+			where: { id },
+			relations: this.findOneRelations
+		});
 	}
 
 	async deleteHall(id: string): Promise<string> {
+		const orders: ActiveOrderEntity[] = await this._ordersRepository.find({
+			where: {
+				table: {
+					hall: { id }
+				}
+			},
+			relations: ["table", "table.hall"]
+		});
+
+		const isActiveOrdersPresent = orders.some((el) => el.status !== OrderStatusEnum.CLOSED);
+
+		if (isActiveOrdersPresent) {
+			await this._hallsRepository.save({ id, isHide: true });
+
+			return `Hall is using in active order(s). Hall will be hide for now`;
+		}
+
 		await this._hallsRepository.delete(id);
 		return `${id} deleted`;
 	}
