+++ src/app/places/services/places.service.ts
@@ -1,36 +1,98 @@
 import { Injectable } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
-import { Repository } from "typeorm";
 
 import { CompanyEntity } from "../../companies/entities";
-import { getFiltersByUrl, getFindOptionsByFilters } from "../../shared/crud";
+import { getFindOptionsByFilters } from "../../shared";
 import type { PaginationArgsDto } from "../../shared/dtos";
-import type { CreatePlaceDto, UpdatePlaceDto } from "../dtos";
+import type { FiltersArgsDto } from "../../shared/dtos";
+import type { PlaceVerificationStatusEnum } from "../../shared/enums";
+import { OrderStatusEnum, UserRoleEnum } from "../../shared/enums";
+import type { IUser } from "../../shared/interfaces";
+import { UserEntity } from "../../users/entities";
+import type { CreatePlaceInput, UpdatePlaceInput } from "../dtos";
+import type { AddEmployeeInput } from "../dtos/add-employee.dto";
 import { PlaceEntity } from "../entities";
 
 @Injectable()
 export class PlacesService {
-	private findRelations = ["company", "company.owner", "halls", "file"];
-	private findOneRelations = ["company", "company.owner", "halls", "file"];
+	private findRelations = [
+		"company",
+		"company.owner",
+		"employees",
+		"halls",
+		"file",
+		// "guests",
+		"paymentSystems",
+		"paymentSystems.paymentSystem"
+	];
+
+	private findOneRelations = [
+		"company",
+		"company.owner",
+		"employees",
+		"halls",
+		"file",
+		// "guests",
+		"paymentSystems",
+		"paymentSystems.paymentSystem"
+	];
 
 	constructor(
-		@InjectRepository(PlaceEntity) private readonly _placesRepository: Repository<PlaceEntity>,
-		@InjectRepository(CompanyEntity) private readonly _companiesRepository: Repository<CompanyEntity>
+		@InjectRepository(PlaceEntity) private readonly _placesRepository,
+		@InjectRepository(CompanyEntity) private readonly _companiesRepository,
+		@InjectRepository(UserEntity) private readonly _usersRepository
 	) {}
 
-	async getPlace(id: string) {
+	async getPlace(id: string, filtersArgs?: FiltersArgsDto[]) {
+		const findOptions = filtersArgs?.length > 0 ? getFindOptionsByFilters(filtersArgs) : ([] as any);
+
 		return this._placesRepository.findOne({
-			where: { id },
+			where: {
+				id,
+				...findOptions.where
+			},
 			relations: this.findOneRelations
 		});
 	}
 
-	async getPlaces({ take, skip, filtersString }: PaginationArgsDto) {
-		const filters = getFiltersByUrl(filtersString);
-		const findOptions = getFindOptionsByFilters(filters) as any;
+	async getPlaceGuests(id: string, { take, skip, filtersArgs }: PaginationArgsDto) {
+		const findOptions = filtersArgs?.length > 0 ? getFindOptionsByFilters(filtersArgs) : ([] as any);
+
+		const [data, count] = await this._usersRepository.findAndCount({
+			where: {
+				placesGuest: {
+					id
+				},
+				...findOptions.where
+			},
+			relations: ["place", "placesGuest"],
+			take,
+			skip
+		});
+
+		return {
+			data,
+			totalCount: count,
+			page: skip / take + 1
+		};
+	}
+
+	async getPlaces({ take, skip, filtersArgs }: PaginationArgsDto, user: IUser) {
+		const findOptions = getFindOptionsByFilters(filtersArgs) as any;
 
 		const [data, count] = await this._placesRepository.findAndCount({
-			where: findOptions.where,
+			where: {
+				...findOptions.where,
+				...(user.role !== UserRoleEnum.ADMIN && user.role !== UserRoleEnum.CLIENT
+					? {
+							company: {
+								owner: {
+									id: user.id
+								}
+							}
+					  }
+					: {})
+			},
 			relations: this.findRelations,
 			take,
 			skip
@@ -43,18 +105,70 @@ export class PlacesService {
 		};
 	}
 
-	async createPlace(place: CreatePlaceDto): Promise<PlaceEntity> {
+	async createPlace(place: CreatePlaceInput): Promise<PlaceEntity> {
 		const savedPlace = await this._placesRepository.save({ ...place, company: { id: place.company } });
 
 		return this._placesRepository.findOne({ where: { id: savedPlace.id } });
 	}
 
-	async updatePlace(id: string, place: UpdatePlaceDto): Promise<PlaceEntity> {
-		return this._placesRepository.save({ id, ...place });
+	async updatePlace(id: string, place: UpdatePlaceInput): Promise<PlaceEntity> {
+		await this._placesRepository.save({ id, ...place });
+
+		return this._placesRepository.findOne({ where: { id }, relations: this.findOneRelations });
 	}
 
 	async deletePlace(id: string): Promise<string> {
+		const place: PlaceEntity = await this._placesRepository.findOne({
+			where: { id },
+			relations: [...this.findOneRelations, "orders"]
+		});
+
+		const isActiveOrdersPresent = place.orders.some((el) => el.status !== OrderStatusEnum.CLOSED);
+
+		if (isActiveOrdersPresent) {
+			await this._placesRepository.save({ ...place, isHide: true });
+
+			return `${place.name} is having active order(s). Place will be hide for now`;
+		}
+
 		await this._placesRepository.delete(id);
 		return `${id} deleted`;
+	}
+
+	async addEmployeeToPlace(employee: AddEmployeeInput) {
+		const place = await this._placesRepository.findOne({
+			where: {
+				id: employee.placeId
+			},
+			relations: ["employees"]
+		});
+
+		const user = await this._usersRepository.findOne({ where: { id: employee.userId } });
+
+		return this._placesRepository.save({ ...place, employees: [...(place.employees || []), { ...user }] });
+	}
+
+	async removeEmployeeFromPlace(employee: AddEmployeeInput) {
+		const place = await this._placesRepository.findOne({
+			where: {
+				id: employee.placeId
+			},
+			relations: ["employees"]
+		});
+
+		return this._placesRepository.save({
+			...place,
+			employees: place.employees.filter((el) => el.id !== employee.userId)
+		});
+	}
+
+	async updatePlaceVerification(placeId: string, status: PlaceVerificationStatusEnum) {
+		const place: PlaceEntity = await this._placesRepository.findOne({
+			where: {
+				id: placeId
+			}
+		});
+
+		return this._placesRepository.save({ ...place, verificationStatus: status });
 	}
 }
